API

Measurements

Measurement interface

The methods in this section forms the interface for measurements. The user can implement their own measurements by following the interface.

QuantumMeasurements.ProportionalMeasurement — Method

ProportionalMeasurement(itr)

Constructs a proportional measurement from the iterator itr.

This measurement is proportional in the sense that, given Given μ = ProportionalMeasurement(itr) and a Bloch vector θ, the probabilities of the outcomes are proportional to μ.measurement_matrix * θ

QuantumMeasurements.empty_measurement — Method

empty_measurement(num_outcomes, dim, T)

Returns an empty measurement matrix with num_outcomes outcomes and dimension dim.

The type T should be either an AbstractMatrix or a ProportionalMeasurement(@ref).

QuantumMeasurements.filter_measurement — Method

filter_measurement(μ, J)

Filters the measurement μ so that only the outcomes J are represented.

QuantumMeasurements.get_dim — Method

get_dim(μ)

Returns the dimension of the measurement μ.

QuantumMeasurements.get_measurement_type — Method

get_measurement_type(μ)

Returns the type of the measurement μ.

QuantumMeasurements.get_num_outcomes — Method

get_num_outcomes(μ)

Returns the number of possible outcomes for a single trial of measurement μ.

QuantumMeasurements.get_probabilities! — Method

get_probabilities!(dest, μ, θ)

Returns the probabilities of the outcomes of the measurement μ given the coefficients θ. The result is stored in dest.

QuantumMeasurements.get_probabilities — Method

get_probabilities(μ, θ)

Returns the probabilities of the outcomes of the measurement μ given the coefficients θ.

QuantumMeasurements.get_trace_part — Method

get_trace_part(μ)

Returns the trace part of the measurement μ.

QuantumMeasurements.get_traceless_part — Method

get_traceless_part(μ)

Returns the traceless part of the measurement μ.

QuantumMeasurements.multithreaded_update_measurement! — Method

multithreaded_update_measurement!(μ, buffers, itr, pars, f!)

Same as update_measurement!(@ref), but uses multithreading.

QuantumMeasurements.post_estimation_routine! — Method

post_estimation_routine!(ρ, θ, μ)

Performs a post-estimation routine on the state ρ, θ for the measurement μ. Default implementation does nothing.

QuantumMeasurements.update_measurement! — Method

update_measurement!(μ, itr)
update_measurement!(μ::AbstractMatrix, buffer, itr, pars, f!)

Updates the measurement μ with an iterator itr. The second sigature is used when the measurement is updated with a buffer and a function f!, which is used to update the buffer. f! should have the signature f!(buffer, pars).

If get_measurement_type, get_num_outcomes and get_probabilities! are implemented, one has also access to get_probabilities:

QuantumMeasurements.get_probabilities — Function

get_probabilities(μ, θ)

Returns the probabilities of the outcomes of the measurement μ given the coefficients θ.

Measurement matrix

The simplest measurement type is just a matrix:

QuantumMeasurements.assemble_measurement_matrix — Method

assemble_measurement_matrix(itr)

Assembles a measurement matrix from an iterator itr. Each row of the matrix is a vectorization of the corresponding element of itr. vectorization is implemented when the element of itr is a matrix or a vector, representing a Positive Operator Valued Measure (POVM) or a Projection Valued Measure (PVM) respectively. Given μ = assemble_measurement_matrix(itr) and a Bloch vector θ, the probabilities of the outcomes are calculated by μ * θ

Proportional Measurements

QuantumMeasurements.ProportionalMeasurement — Type

ProportionalMeasurement(itr)

Constructs a proportional measurement from the iterator itr.

This measurement is proportional in the sense that, given Given μ = ProportionalMeasurement(itr) and a Bloch vector θ, the probabilities of the outcomes are proportional to μ.measurement_matrix * θ

Tomography Methods

This package contains several methods for quantum state tomography. They should be passed to the following function in order to perform the tomography:

QuantumMeasurements.estimate_state — Function

estimate_state(outcomes, μ, method; kwargs...)

Estimate the state of a quantum system given the outcomes of a set of measurements represented by μ using the specified method.

If outcomes is an array of integers, it is interpreted as the number of times each measurement outcome was observed. If outcomes is an array of floats, it is interpreted as the frequency of each measurement outcome, so the sum of the elements must be 1.

Outputs a tuple (ρ, θ) where ρ is the estimated density matrix and θ is the corresponding Bloch vector. Additionally, the BayesianInference method also returns a covariance matrix Σ representing the uncertainty in the estimated Bloch vector.

For AbstractLinearInversion (LinearInversion, PreAllocatedLinearInversion and NormalEquations) methods, there are no additional keyword arguments.

For the MaximumLikelihood method, the following keyword arguments are available:

x₀=zeros(get_measurement_type(μ), get_dim(μ)^2 - 1): Initial guess for the Bloch vector. Default is a maximally mixed state.

t=0.4: Initial step size for the backtracking line search.

β=0.8: Factor for the backtracking line search.

max_iter=10^3: Maximum number of iterations.

tol=1e-10: Tolerance for the convergence criterion.

For the BayesianInference method, the following keyword arguments are available:

verbose=false: whether to print information about the run.

σ=get_measurement_type(μ)(1e-2): the initial standard deviation of the proposal distribution.

log_prior=θ -> zero(get_measurement_type(μ)): the log-prior function.

θ₀=zeros(get_measurement_type(μ), get_dim(μ)^2 - 1): the initial state of the chain.

nsamples=10^4: the number of samples to take.

nwarm=10^3: the number of warm-up samples to take.

chain=nothing: if not nothing, store the chain in this matrix.

The available methods are:

QuantumMeasurements.LinearInversion — Type

LinearInversion <: AbstractLinearInversion

A type representing a linear inversion method for estimating the state of a quantum system. The equations are solved by a QR decomposition (\ operator).

QuantumMeasurements.PreAllocatedLinearInversion — Type

PreAllocatedLinearInversion(μ) <: AbstractLinearInversion

Create a PreAllocatedLinearInversion instance for the given measurement μ.

This method precomputes the pseudo-inverse of the corresponding measurement matrix, so that subsequent calls to estimate_state are faster.

QuantumMeasurements.NormalEquations — Type

NormalEquations(μ) <: AbstractLinearInversion

Create a NormalEquations instance for the given measurement μ. This method solves the normal equations for the given measurement matrix. This can be faster for systems with a large number of measurements.

QuantumMeasurements.MaximumLikelihood — Type

MaximumLikelihood

A type used to specify the maximum likelihood estimation method.

QuantumMeasurements.BayesianInference — Type

BayesianInference

Create a Bayesian inference instance.

This method is passed to the estimate_state method in order to perform quantum state tomography.

Gell-Mann Matrices

QuantumMeasurements.GellMannMatrices — Type

GellMannMatrices(dim, ::Type{T}=ComplexF32) where {T<:Complex}

Iterator over the generalized Gell-Mann matrices of dimension dim with elements of type T.

Vectorization

QuantumMeasurements.traceless_vectorization — Function

traceless_vectorization(A)

Returns a vector θ containing the projections of A onto the GellMannMatrices. A can be a matrix or a vector. In the latter case, it is treated as a projector. The vector θ has length dim^2 - 1, where dim is the dimension of A.

QuantumMeasurements.traceless_vectorization! — Function

traceless_vectorization!(θ, A)

Same as traceless_vectorization, but stores the result in θ.

QuantumMeasurements.density_matrix_reconstruction — Function

density_matrix_reconstruction(θ)

Returns the density matrix corresponding to the Bloch vector θ.

QuantumMeasurements.density_matrix_reconstruction! — Function

density_matrix_reconstruction!(ρ, θ)

Same as density_matrix_reconstruction, but stores the result in ρ.

Generation of Random States

QuantumMeasurements.GinibreEnsamble — Type

GinibreEnsamble(dim::Int)

A type representing a Ginibre ensamble of complex matrices of dimension dim.

QuantumMeasurements.HaarUnitary — Type

HaarUnitary(dim::Int)

A type representing a Haar-random unitary matrix of dimension dim.

QuantumMeasurements.HaarVector — Type

HaarVector(dim::Int)

A type representing a Haar-random unitary vector of dimension dim.

QuantumMeasurements.ProductMeasure — Type

ProductMeasure(dim::Int)

A type representing a measure on the density states. It is a product Haar measure on the unitary group and a uniform (Lebesgue) measure on the simplex.

QuantumMeasurements.Simplex — Type

Simplex(dim::Int)

A type representing a random point on the simplex embeded in a space of dimension dim.

Base.rand — Method

Base.rand([rng = default_rng()], type, [n_samples])

Sample n_samples from type.

If n_samples is not provided, a single sample is returned.

Possible values for type are HaarUnitary, HaarVector, Simplex, ProductMeasure, and GinibreEnsamble.

Misc

LinearAlgebra.isposdef! — Method

isposdef!(ρ, xs, set)

Calculate the linear combination of the elements of set with the coefficients xs and check if the result is a positive definite matrix.

QuantumMeasurements.fidelity — Method

fidelity(ρ, σ)

Calculate the fidelity between two quantum states.

Either state can be pure or mixed, and represented by vectors or matrices, respectively.

QuantumMeasurements.fisher — Method

fisher(μ, θs)

Returns the Fisher information matrix for the measurement μ and the state decribed by a Bloch vector θs.

QuantumMeasurements.polarization_state — Method

polarization_state(s::Symbol, ::Type{T}=ComplexF32) where {S, T}

Return the polarization state corresponding to the symbol S as a vector of type T. S can be one of the following symbols:

:H for horizontal polarization

:V for vertical polarization

:D for diagonal polarization

:A for antidiagonal polarization

:R for right-handed circular polarization

:L for left-handed circular polarization

QuantumMeasurements.project2density! — Method

project2density(ρ)

Project a Hermitian matrix ρ to a density matrix by setting the negative eigenvalues to zero and normalizing the trace to 1.

QuantumMeasurements.project2pure — Method

project2pure(ρ)

Project a Hermitian matrix ρ to a pure state by returning the eigenvector corresponding to the largest eigenvalue.

QuantumMeasurements.simulate_outcomes — Method

simulate_outcomes([rng,] state, measurement, N; atol=1e-4)

Simulate the outcomes of a measurement measurement on a state state N times. The rng parameter is a random number generator. state can be either a vector, describing a pure state, or a matrix, describing a mixed state. atol is the absolute tolerance for the sum of the probabilities to be 1. If the tolerance is not met, an error is thrown, which probably means that there is something wrong with the provided state or measurement.